refactor: Queue 프로토콜을 생성하여, CustomerQueue 에 적용
CustomerQueue에 들어가는 Customer타입 생성
