+++ BankManagerConsoleApp/BankManagerConsoleApp/Model/Customer.swift
@@ -1,0 +1,12 @@
+//
+//  Customer.swift
+//  BankManagerConsoleApp
+//
+//  Created by 이원빈 on 2022/06/30.
+//
+
+import Foundation
+
+struct Customer: Equatable {
+    var number: Int
+}
