+++ BankManagerConsoleApp/BankManagerConsoleApp/Model/CustomerQueue.swift
@@ -7,35 +7,13 @@
 
 import Foundation
 
-struct CustomerQueue<T> {
-    private(set) var queue: LinkedList<T>
-    var isEmpty: Bool {
-        queue.isEmpty
-    }
-    var peek: T? {
-        queue.peek
-    }
-    var count: Int {
-        queue.count
-    }
+struct CustomerQueue<Customer>: Queue {
     
-    func enqueue(data: T) {
-        queue.append(data: data)
-    }
-    
-    @discardableResult
-    func dequeue() -> T? {
-        queue.removeFirst()
-    }
-    
-    func clear() {
-        queue.clear()
-    }
-    
-    init(elements: [T] = []) {
+    private(set) var queue: LinkedList<Customer>
+    init(elements: [Customer] = []) {
         queue = LinkedList()
         elements.forEach {
-            enqueue(data: $0)
+            self.enqueue(data: $0)
         }
     }
 }
